Move raw tab params into separate file
