+++ rtengine/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(PROCPARAM_SRCS
     procparams.cc
     params/locallab.cc
+    params/raw.cc
     params/serdes.cc
 )
